requirements.txt
torch>=2.0.0
transformers>=4.30.0
langdetect>=1.0.9
google-generativeai>=0.3.2
PyQt5>=5.15.9
scikit-learn>=1.1.3
protobuf>=4.21.0
